Move models to discovery and collection modules
