+++ crates/karva_core/src/discovery/mod.rs
@@ -1,5 +1,11 @@
 pub mod discoverer;
+pub mod models;
 pub mod visitor;
 
 pub use discoverer::Discoverer;
+pub use models::{
+    function::{TestFunction, TestFunctionDisplay},
+    module::{DiscoveredModule, ModuleType, StringModule},
+    package::{DiscoveredPackage, StringPackage},
+};
 pub use visitor::{FunctionDefinitionVisitor, discover};
