+++ crates/karva_core/src/fixture/manager.rs
@@ -3,8 +3,8 @@ use std::collections::HashMap;
 use pyo3::{prelude::*, types::PyAny};
 
 use crate::{
+    discovery::DiscoveredPackage,
     fixture::{Finalizer, Finalizers, Fixture, FixtureScope, HasFixtures, RequiresFixtures},
-    models::DiscoveredPackage,
     utils::partition_iter,
 };
