--- crates/karva_core/src/models/mod.rs
@@ -1,9 +1,0 @@
-pub mod module;
-pub mod package;
-pub mod test_case;
-pub mod test_function;
-
-pub use module::{DiscoveredModule, ModuleType, StringModule};
-pub use package::{DiscoveredPackage, StringPackage};
-pub use test_case::TestCase;
-pub use test_function::TestFunction;
